+++ src/main/java/com/hubspot/jinjava/doc/JinjavaDocExpTest.java
@@ -1,9 +1,11 @@
 package com.hubspot.jinjava.doc;
 
+import java.util.Map;
+
 public class JinjavaDocExpTest extends JinjavaDocItem {
 
-  public JinjavaDocExpTest(String name, String desc, String aliasOf, JinjavaDocParam[] params, JinjavaDocSnippet[] snippets) {
-    super(name, desc, aliasOf, params, snippets);
+  public JinjavaDocExpTest(String name, String desc, String aliasOf, boolean deprecated, JinjavaDocParam[] params, JinjavaDocSnippet[] snippets, Map<String, String> meta) {
+    super(name, desc, aliasOf, deprecated, params, snippets, meta);
   }
 
 }