+++ src/main/java/com/hubspot/jinjava/doc/JinjavaDocItem.java
@@ -1,19 +1,25 @@
 package com.hubspot.jinjava.doc;
 
+import java.util.Map;
+
 public abstract class JinjavaDocItem {
 
   private final String name;
   private final String desc;
   private final String aliasOf;
+  private final boolean deprecated;
   private final JinjavaDocParam[] params;
   private final JinjavaDocSnippet[] snippets;
+  private final Map<String, String> meta;
   
-  public JinjavaDocItem(String name, String desc, String aliasOf, JinjavaDocParam[] params, JinjavaDocSnippet[] snippets) {
+  public JinjavaDocItem(String name, String desc, String aliasOf, boolean deprecated, JinjavaDocParam[] params, JinjavaDocSnippet[] snippets, Map<String, String> meta) {
     this.name = name;
     this.desc = desc;
     this.aliasOf = aliasOf;
+    this.deprecated = deprecated;
     this.params = params;
     this.snippets = snippets;
+    this.meta = meta;
   }
 
   public String getName() {
@@ -28,12 +34,20 @@ public abstract class JinjavaDocItem {
     return aliasOf;
   }
   
+  public boolean isDeprecated() {
+    return deprecated;
+  }
+  
   public JinjavaDocParam[] getParams() {
     return params;
   }
   
   public JinjavaDocSnippet[] getSnippets() {
     return snippets;
+  }
+  
+  public Map<String, String> getMeta() {
+    return meta;
   }
   
 }