+++ src/main/java/com/hubspot/jinjava/doc/JinjavaDocTag.java
@@ -1,11 +1,13 @@
 package com.hubspot.jinjava.doc;
 
+import java.util.Map;
+
 public class JinjavaDocTag extends JinjavaDocItem {
 
   private final boolean empty;
   
-  public JinjavaDocTag(String name, boolean empty, String desc, String aliasOf, JinjavaDocParam[] params, JinjavaDocSnippet[] snippets) {
-    super(name, desc, aliasOf, params, snippets);
+  public JinjavaDocTag(String name, boolean empty, String desc, String aliasOf, boolean deprecated, JinjavaDocParam[] params, JinjavaDocSnippet[] snippets, Map<String, String> meta) {
+    super(name, desc, aliasOf, deprecated, params, snippets, meta);
     this.empty = empty;
   }